move: 영역 별로 폴더 분리하기(관리자 페이지)
