+++ src/refactoring/components/CartPage/cartSection/CartList.tsx
@@ -1,6 +1,6 @@
-import type { CartItem } from '../../../types';
-import CartListItem from './CartListItem';
-import { getAppliedDiscount } from '../../utils/discountUtils.ts';
+import type { CartItem } from '../../../../types.ts';
+import CartListItem from './CartListItem.tsx';
+import { getAppliedDiscount } from '../../../utils/discountUtils.ts';
 
 interface CartListProps {
   cart: CartItem[];
