+++ src/refactoring/components/CartPage/productSection/ProductListItem.tsx
@@ -1,7 +1,7 @@
 import { useCallback, memo } from 'react';
 
-import { getMaxDiscount } from '../../utils/discountUtils.ts';
-import type { Product } from '../../../types';
+import { getMaxDiscount } from '../../../utils/discountUtils.ts';
+import type { Product } from '../../../../types.ts';
 
 interface ProductListItemProps {
   product: Product;
